+++ test-d/entrypoint-browser.test-d.ts
@@ -144,6 +144,10 @@ BeforeAll(function () {
   expectType<Mocha.Context>(this);
 });
 
+AfterAll(function () {
+  expectType<Mocha.Context>(this);
+});
+
 Before(function () {
   expectType<Mocha.Context>(this);
 });
@@ -165,10 +169,6 @@ After({}, function () {
 });
 
 After({ tags: "foo", name: "bar" }, function () {
-  expectType<Mocha.Context>(this);
-});
-
-AfterAll(function () {
   expectType<Mocha.Context>(this);
 });